# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StePanel API
  version: 0.6.0
  description: |
    Authenticated server operations, service inventory, and migration API.
    Mutating requests require the `stepanel_session` cookie and the matching
    `stepanel_csrf` value in the `X-CSRF-Token` header or `csrf` form field.
    Responses include an `X-Request-ID` correlation identifier.
servers:
  - url: http://127.0.0.1:8090
security:
  - sessionCookie: []
paths:
...
  /api/databases:
    get:
      summary: List site-owned databases managed by StePanel
      responses:
        '200': {description: Database, owner, user, encoding, and allocated bytes}
    post:
      summary: Create a database and least-privilege site user
      responses:
        '201': {description: Database and user created; password is never returned or stored by StePanel}
        '409': {description: Object already exists or provisioning failed safely}
  /api/databases/{name}:
    get:
      summary: Inspect one managed database without exposing credentials
      parameters:
        - {name: name, in: path, required: true, schema: {type: string, pattern: '^[a-z0-9_]{1,63}$'}}
      responses:
        '200': {description: Database owner, user, encoding, and allocated bytes}
        '404': {description: Managed database not found}
    delete:
      summary: Create a safety dump, then delete a managed database and user
      parameters:
        - {name: name, in: path, required: true, schema: {type: string, pattern: '^[a-z0-9_]{1,63}$'}}
      responses:
        '200': {description: Database deleted with safety-backup path and checksum}
        '503': {description: Deletion refused because the safety backup failed}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: stepanel_session